+++ src/main/java/dev/muon/dynamic_difficulty/data/DimensionsLevelingSettingsReloader.java
@@ -83,16 +83,7 @@ public class DimensionsLevelingSettingsReloader extends ContextAwareReloadListen
   }
 
   private static DimensionLevelingSettings createDefaultSettings() {
-    return new DimensionLevelingSettings(
-        Config.COMMON.startingLevel.get(),
-        Config.COMMON.maxLevel.get(),
-        Config.COMMON.levelsPerDistance.get().floatValue(),
-        Config.COMMON.levelsPerDeepness.get().floatValue(),
-        Config.COMMON.randomLevelBonus.get(),
-        null,
-        64, // Default sea level
-        Config.COMMON.levelsPerHeight.get().floatValue(),
-        Map.of());
+    return DimensionLevelingSettings.createDefault();
   }
 
   @Override